v3.15.0

+++ src/main/java/com/example/beautysalonRESTAPI/dto/employees/skillsDTO.java
@@ -2,18 +2,18 @@ package com.example.beautysalonRESTAPI.dto.employees;
 
 public class skillsDTO {
     
-    private int id_employee;
-    private int id_service;
-    public int getId_employee() {
+    private Long id_employee;
+    private Long id_service;
+    public Long getId_employee() {
         return id_employee;
     }
-    public void setId_employee(int id_employee) {
+    public void setId_employee(Long id_employee) {
         this.id_employee = id_employee;
     }
-    public int getId_service() {
+    public Long getId_service() {
         return id_service;
     }
-    public void setId_service(int id_service) {
+    public void setId_service(Long id_service) {
         this.id_service = id_service;
     }
 }
